Move thrift utils only used by scrooge to contrib/scrooge.
Since the top-level utility functions are moved out of thrift_util,
this change takes the opportunity to also rename thrift_util.py and
likewise its dependency, binary_util.py.  The rest is import and BUILD
dep fallout.
