+++ contrib/scrooge/src/python/pants/contrib/scrooge/tasks/BUILD
@@ -2,19 +2,20 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_library(
-  name = 'tasks',
-  dependencies = [
+  name='tasks',
+  dependencies=[
     ':scrooge_gen',
     ':thrift_linter',
   ],
 )
 
 python_library(
-  name = 'scrooge_gen',
-  sources = ['scrooge_gen.py'],
-  dependencies = [
-    ':java_thrift_library_fingerprint_strategy',
+  name='scrooge_gen',
+  sources=['scrooge_gen.py'],
+  dependencies=[
     '3rdparty/python/twitter/commons:twitter.common.collections',
+    ':java_thrift_library_fingerprint_strategy',
+    ':thrift_util',
     'src/python/pants/backend/codegen/subsystems:thrift_defaults',
     'src/python/pants/backend/codegen/targets:java',
     'src/python/pants/backend/jvm/targets:java',
@@ -27,14 +28,13 @@ python_library(
     'src/python/pants/base:exceptions',
     'src/python/pants/option',
     'src/python/pants/util:dirutil',
-    'src/python/pants:thrift_util',
   ],
 )
 
 python_library(
-  name = 'thrift_linter',
-  sources = ['thrift_linter.py'],
-  dependencies = [
+  name='thrift_linter',
+  sources=['thrift_linter.py'],
+  dependencies=[
     'src/python/pants/backend/jvm/tasks:jvm_tool_task_mixin',
     'src/python/pants/backend/jvm/tasks:nailgun_task',
     'src/python/pants/base:exceptions',
@@ -44,9 +44,14 @@ python_library(
 )
 
 python_library(
-  name = 'java_thrift_library_fingerprint_strategy',
-  sources = ['java_thrift_library_fingerprint_strategy.py'],
-  dependencies = [
+  name='thrift_util',
+  sources=['thrift_util.py'],
+)
+
+python_library(
+  name='java_thrift_library_fingerprint_strategy',
+  sources=['java_thrift_library_fingerprint_strategy.py'],
+  dependencies=[
     'src/python/pants/backend/codegen/targets:java',
     'src/python/pants/base:fingerprint_strategy',
   ],
