+++ contrib/scrooge/tests/python/pants_test/contrib/scrooge/tasks/BUILD
@@ -2,18 +2,30 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 target(
-  name = 'tasks',
-  dependencies = [
+  name='tasks',
+  dependencies=[
+    ':java_thrift_library_fingerprint_strategy',
     ':scrooge_gen',
     ':thrift_linter_integration',
-    ':java_thrift_library_fingerprint_strategy',
+    ':thrift_util',
   ],
 )
 
 python_tests(
-  name = 'scrooge_gen',
-  sources = ['test_scrooge_gen.py'],
-  dependencies = [
+  name='java_thrift_library_fingerprint_strategy',
+  sources=['test_java_thrift_library_fingerprint_strategy.py'],
+  dependencies=[
+    'contrib/scrooge/src/python/pants/contrib/scrooge/tasks:java_thrift_library_fingerprint_strategy',
+    'src/python/pants/backend/codegen/subsystems:thrift_defaults',
+    'tests/python/pants_test/base:context_utils',
+    'tests/python/pants_test/tasks:task_test_base',
+  ]
+)
+
+python_tests(
+  name='scrooge_gen',
+  sources=['test_scrooge_gen.py'],
+  dependencies=[
     '3rdparty/python:mock',
     'contrib/scrooge/src/python/pants/contrib/scrooge/tasks:scrooge_gen',
     'src/python/pants/backend/codegen/targets:java',
@@ -25,20 +37,20 @@ python_tests(
 )
 
 python_tests(
-  name = 'thrift_linter_integration',
-  sources = ['test_thrift_linter_integration.py'],
-  dependencies = [
+  name='thrift_linter_integration',
+  sources=['test_thrift_linter_integration.py'],
+  dependencies=[
     'tests/python/pants_test:int-test',
   ],
 )
 
 python_tests(
-  name = 'java_thrift_library_fingerprint_strategy',
-  sources = ['test_java_thrift_library_fingerprint_strategy.py'],
-  dependencies = [
-    'contrib/scrooge/src/python/pants/contrib/scrooge/tasks:java_thrift_library_fingerprint_strategy',
-    'src/python/pants/backend/codegen/subsystems:thrift_defaults',
-    'tests/python/pants_test/base:context_utils',
-    'tests/python/pants_test/tasks:task_test_base',
+  name='thrift_util',
+  sources=['test_thrift_util.py'],
+  dependencies=[
+    'contrib/scrooge/src/python/pants/contrib/scrooge/tasks:thrift_util',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:dirutil',
+    'tests/python/pants_test:base_test',
   ]
 )
