+++ contrib/scrooge/tests/python/pants_test/contrib/scrooge/tasks/test_thrift_util.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
+# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
@@ -8,9 +8,10 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import os
 import unittest
 
-from pants.thrift_util import find_includes, find_root_thrifts
 from pants.util.contextutil import temporary_dir
 from pants.util.dirutil import safe_open
+
+from pants.contrib.scrooge.tasks.thrift_util import find_includes, find_root_thrifts
 
 
 class ThriftUtilTest(unittest.TestCase):
@@ -35,8 +36,8 @@ class ThriftUtilTest(unittest.TestCase):
       b_included = self.write(os.path.join(b, 'b_included.thrift'), '# noop')
       c_included = self.write(os.path.join(b, 'c_included.thrift'), '# noop')
 
-      self.assertEquals(set([a_included, b_included, c_included]),
-                        find_includes(basedirs=set([a, b]), source=main))
+      self.assertEquals({a_included, b_included, c_included},
+                        find_includes(basedirs={a, b}, source=main))
 
   def test_find_includes_exception(self):
     with temporary_dir() as dir:
@@ -47,13 +48,13 @@ class ThriftUtilTest(unittest.TestCase):
         include "b_included.thrift"
       ''')
       self.write(os.path.join(a, 'sub', 'a_included.thrift'), '# noop')
-      self.assertRaises(ValueError, find_includes, basedirs=set([a]), source=main)
+      self.assertRaises(ValueError, find_includes, basedirs={a}, source=main)
 
   def test_find_root_thrifts(self):
     with temporary_dir() as dir:
       root_1 = self.write(os.path.join(dir, 'root_1.thrift'), '# noop')
       root_2 = self.write(os.path.join(dir, 'root_2.thrift'), '# noop')
-      self.assertEquals(set([root_1, root_2]),
+      self.assertEquals({root_1, root_2},
                         find_root_thrifts(basedirs=[], sources=[root_1, root_2]))
 
     with temporary_dir() as dir:
@@ -61,4 +62,4 @@ class ThriftUtilTest(unittest.TestCase):
       self.write(os.path.join(dir, 'mid_1.thrift'), 'include "leaf_1.thrift"')
       self.write(os.path.join(dir, 'leaf_1.thrift'), '# noop')
       root_2 = self.write(os.path.join(dir, 'root_2.thrift'), 'include "root_1.thrift"')
-      self.assertEquals(set([root_2]), find_root_thrifts(basedirs=[], sources=[root_1, root_2]))
+      self.assertEquals({root_2}, find_root_thrifts(basedirs=[], sources=[root_1, root_2]))
