+++ src/python/pants/backend/core/tasks/BUILD
@@ -94,10 +94,11 @@ python_library(
   name = 'confluence_publish',
   sources = ['confluence_publish.py'],
   dependencies = [
-    ':common',
     '3rdparty/python/twitter/commons:twitter.common.confluence',
+    ':task',
     'src/python/pants/backend/core/targets:common',
-    'src/python/pants:binary_util',
+    'src/python/pants/base:exceptions',
+    'src/python/pants/binaries:binary_util',
     'src/python/pants/util:dirutil',
   ],
 )
@@ -201,8 +202,8 @@ python_library(
     'src/python/pants/base:exceptions',
     'src/python/pants/base:generator',
     'src/python/pants/base:workunit',
+    'src/python/pants/binaries:binary_util',
     'src/python/pants/util:dirutil',
-    'src/python/pants:binary_util',
   ],
 )
 
@@ -246,8 +247,8 @@ python_library(
     ':task',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:run_info',
+    'src/python/pants/binaries:binary_util',
     'src/python/pants/reporting',
-    'src/python/pants:binary_util',
   ],
 )
