+++ src/python/pants/backend/jvm/tasks/ivy_resolve.py
@@ -11,13 +11,13 @@ import time
 from collections import defaultdict
 from textwrap import dedent
 
-from pants import binary_util
 from pants.backend.jvm.ivy_utils import IvyUtils
 from pants.backend.jvm.targets.jar_library import JarLibrary
 from pants.backend.jvm.tasks.ivy_task_mixin import IvyTaskMixin
 from pants.backend.jvm.tasks.nailgun_task import NailgunTask
 from pants.base.cache_manager import VersionedTargetSet
 from pants.base.exceptions import TaskError
+from pants.binaries import binary_util
 from pants.goal.products import UnionProducts
 from pants.ivy.ivy_subsystem import IvySubsystem
 from pants.util.dirutil import safe_mkdir
