+++ src/python/pants/backend/project_info/tasks/BUILD
@@ -102,8 +102,8 @@ python_library(
   name = 'ide_gen',
   sources = ['ide_gen.py'],
   dependencies = [
-    ':projectutils',
     '3rdparty/python/twitter/commons:twitter.common.collections',
+    ':projectutils',
     'src/python/pants/backend/core/tasks:task',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:scala',
@@ -111,7 +111,7 @@ python_library(
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:source_root',
-    'src/python/pants:binary_util',
+    'src/python/pants/binaries:binary_util',
     'src/python/pants/util:dirutil',
   ],
 )
