+++ tests/python/pants_test/backend/python/BUILD
@@ -4,14 +4,14 @@
 target(
   name='python',
   dependencies=[
-    ':test_python_requirement_list',
-    ':test_python_chroot',
+    ':python_chroot',
+    ':python_requirement_list',
     'tests/python/pants_test/backend/python/tasks'
   ]
 )
 
 python_tests(
-  name='test_python_chroot',
+  name='python_chroot',
   sources=['test_python_chroot.py'],
   dependencies=[
     '3rdparty/python:pex',
@@ -23,21 +23,21 @@ python_tests(
     'src/python/pants/backend/python:python_setup',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:source_root',
+    'src/python/pants/binaries:binary_util',
+    'src/python/pants/binaries:thrift_util',
     'src/python/pants/ivy',
     'src/python/pants/util:contextutil',
-    'src/python/pants:binary_util',
-    'src/python/pants:thrift_util',
     'tests/python/pants_test/base:context_utils',
     'tests/python/pants_test:base_test',
   ]
 )
 
 python_tests(
-  name='test_python_requirement_list',
+  name='python_requirement_list',
   sources=['test_python_requirement_list.py'],
   dependencies=[
-    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/backend/python/targets:python',
+    'src/python/pants/backend/python:python_requirement',
     'src/python/pants/base:build_file_aliases',
     'tests/python/pants_test:base_test'
   ]
