+++ tests/python/pants_test/binaries/BUILD
@@ -1,0 +1,18 @@
+# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+target(
+  name='binaries',
+  dependencies=[
+    ':binary_util',
+  ]
+)
+
+python_tests(
+  name='binary_util',
+  sources=['test_binary_util.py'],
+  dependencies=[
+    'src/python/pants/binaries:binary_util',
+    'tests/python/pants_test:base_test',
+  ]
+)
